requirements.txt
addict==2.4.0
appdirs==1.4.4
asttokens==3.0.0
Bottleneck @ file:///croot/bottleneck_1731058641041/work
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
certifi==2025.8.3
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer==3.4.3
click==8.2.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cramjam==2.11.0
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
docker-pycreds==0.4.0
exceptiongroup==1.3.0
executing==2.2.0
fastparquet==2024.11.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.1.5
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
ipykernel==6.30.0
ipython==8.37.0
jedi==0.19.2
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.3
matplotlib-inline==0.1.7
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numexpr @ file:///croot/numexpr_1752521720723/work
numpy==1.26.4
onnx==1.18.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
packaging==25.0
pandas @ file:///croot/pandas_1752846035684/work/dist/pandas-2.3.1-cp310-cp310-linux_x86_64.whl#sha256=c6d2c323b368a4b551221ea68749d31c7a4ae0426805ee73defef87a55d74119
parso==0.8.4
pexpect==4.9.0
pillow @ file:///croot/pillow_1752524526621/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
prompt_toolkit==3.0.51
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.10
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.19.2
pyparsing==3.2.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz @ file:///croot/pytz_1752135852232/work
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
-e git+https://github.com/ngbrjyj/A2LC.git@c680e4adad70cb1f296458c18184672fea9cf3c8#egg=segment_anything&subdirectory=Grounded-Segment-Anything/segment_anything
sentry-sdk==2.35.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
supervision==0.26.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.17
tokenizers==0.21.2
tomli==2.2.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
urllib3==2.5.0
visdom==0.2.4
wandb==0.16.6
wcwidth==0.2.13
websocket-client==1.8.0
yapf==0.43.0
